Clean up and deduplicate renderer implementations
